chore: refactor concurrent implementations
