+++ Cargo.toml
@@ -45,7 +45,7 @@ name = "store"
 harness = false
 
 [features]
-concurrent = ["dep:rayon"]
+concurrent = ["dep:rayon", "hashbrown?/rayon"]
 default = ["std", "concurrent"]
 executable = ["dep:clap", "dep:rand-utils", "std"]
 smt_hashmaps = ["dep:hashbrown"]
