+++ src/merkle/mod.rs
@@ -23,9 +23,11 @@ pub use path::{MerklePath, RootPath, ValuePath};
 mod smt;
 #[cfg(feature = "internal")]
 pub use smt::build_subtree_for_bench;
+#[cfg(feature = "internal")]
+pub use smt::SubtreeLeaf;
 pub use smt::{
     LeafIndex, MutationSet, SimpleSmt, Smt, SmtLeaf, SmtLeafError, SmtProof, SmtProofError,
-    SubtreeLeaf, SMT_DEPTH, SMT_MAX_DEPTH, SMT_MIN_DEPTH,
+    SMT_DEPTH, SMT_MAX_DEPTH, SMT_MIN_DEPTH,
 };
 
 mod mmr;
